Fix a bug preventing auto-conversion from skip to dimen
When it should have been looking for an "internal" dimension (for
constructions such as "\dimen123=\dimen456") unravel was looking for a
skip.  If the right-hand side was a dimension, all was good, but if
it was a skip (e.g. "\dimen123=\skip456" or "\dimen123=\parskip") then
the whole value of the skip was used: "\dimen123=0pt plus 1pt".  That
left stuff in the input stream, broke "\ifdim", etc.

This commit also fixes a bug where "\prevdepth" was treated as an
integer rather than a dimension in some cases.