Extra test for optional space after fil...l (see #38)
Note that TeX searches for more "l" by repeatedly calling scan_keyword
so "fil \space\space l" is valid and TeX continues deleting spaces
after the last "l" it found (searching for more "l").  To avoid this
and really test for the <optional space> we do "fill\noexpand\space":
then when looking for more "l", TeX finds a \notexpanded:\space, and
when looking for the <optional space> TeX expands and deletes \space.