Account partially for outside changes, more needed
In particular remove uses of deprecated expl3 commands, and update
some test logs to account for changes to the LaTeX kernel.  There
remains a major problem with the use of \tl_set:Nx because this is
now defined in terms of \use:x, which itself uses intermediate
storage that is shared between the unravel package and all other
code.