Switch fully from x-type to e-type expansion
This avoids all issues with the internal variable that is used
in \use:x, \exp_args:Nx and similar.  There remains the risk that
some command we rely on could use \tl_set:Nx under the hood, but
that risk is limited by the push towards e-type expansion in the
kernel as well.